# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 86.88 B&B — REST API
  version: 1.0.0
  description: |
    Backend API for the 86.88 民宿 (B&B) platform.

    Serves as the bridge between the PostgreSQL database and the frontend apps:
    - `apps/website` — public-facing Next.js site (8688bnb.com)
    - `apps/admin` — admin dashboard (admin.8688bnb.com)

    ## Authentication
    Admin authentication uses JWT tokens stored in `__Host-8688_session` HttpOnly cookies.
    All state-changing requests require a CSRF token via the `X-CSRF-Token` header.

    ## Rate Limiting
    - Auth endpoints: 20 requests / 15 minutes
    - General API: 200 requests / 15 minutes
    - Public reads: 500 requests / 15 minutes
    - Webhook ingress: 100 requests / 15 minutes

    Rate limit info is returned in response headers:
    - `X-RateLimit-Limit` — max requests in window
    - `X-RateLimit-Remaining` — requests left
    - `X-RateLimit-Reset` — window reset time (Unix epoch)

  contact:
    name: 8688 BnB Dev
    email: dev@8688bnb.com
  license:
    name: Private
    identifier: UNLICENSED
servers:
  - url: http://localhost:3333/api/v1
    description: Local development
  - url: https://api.8688bnb.com/api/v1
    description: Production (via Cloudflare Tunnel)
...
security: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-8688_session
      description: |
        JWT stored in a hardened HttpOnly cookie.
        Set automatically on successful login.
        Cookie attributes: HttpOnly, Secure, SameSite=Strict, Path=/
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        CSRF token required for all state-changing requests (POST, PUT, DELETE).
        Obtained from the login response or `GET /auth/csrf-token`.
    webhookApiKey:
      type: apiKey
      in: header
      name: X-Webhook-Key
      description: |
        API key for authenticating incoming OTA channel manager webhooks.
        Configured in the WEBHOOK_API_KEY environment variable.

  # ─── Reusable Schemas ────────────────────────────────────────────────────────
  schemas:
...
    Booking:
      type: object
      properties:
        id:
          type: integer
          example: 101
        room_id:
          type: integer
          example: 1
        room:
          $ref: '#/components/schemas/RoomSummary'
        source:
          type: string
          enum: [website, phone, line, ota, walk_in, admin]
          example: website
          description: How this booking was created
        ota_platform:
          type: string
          nullable: true
          example: agoda
          description: OTA platform name if source is "ota"
        ota_booking_id:
          type: string
          nullable: true
          example: AGD-123456
          description: External booking ID from OTA platform
        check_in:
          type: string
          format: date
          example: '2026-07-15'
        check_out:
          type: string
          format: date
          example: '2026-07-17'
        guest_name:
          type: string
          example: 陳小明
        guest_phone:
          type: string
          example: '0912-345-678'
        guest_email:
          type: string
          format: email
          nullable: true
          example: guest@example.com
        guest_count:
          type: integer
          minimum: 1
          example: 2
        total_price:
          type: number
          format: decimal
          nullable: true
          example: 6400
          description: Total price in TWD (calculated or manually set)
        status:
          type: string
          enum: [pending, confirmed, checked_in, checked_out, cancelled, no_show]
          example: confirmed
        notes:
          type: string
          nullable: true
          example: 旅客希望加枕頭
          description: Internal notes visible only to admin
        internal_notes:
          type: array
          items:
            $ref: '#/components/schemas/BookingNote'
          description: Timestamped internal note log
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - room_id
        - source
        - check_in
        - check_out
        - guest_name
        - guest_count
        - status
        - created_at
        - updated_at
    BookingCreateRequest:
      type: object
      description: |
        Used by the public booking form on the website.
        Status defaults to "pending" — boss confirms via admin.
      properties:
        room_id:
          type: integer
          example: 1
        check_in:
          type: string
          format: date
          example: '2026-07-15'
        check_out:
          type: string
          format: date
          example: '2026-07-17'
        guest_name:
          type: string
          minLength: 1
          maxLength: 100
          example: 陳小明
        guest_phone:
          type: string
          minLength: 1
          maxLength: 30
          example: '0912-345-678'
        guest_email:
          type: string
          format: email
          maxLength: 255
        guest_count:
          type: integer
          minimum: 1
          maximum: 10
          example: 2
        notes:
          type: string
          maxLength: 1000
          description: Guest message / special requests
      required: [room_id, check_in, check_out, guest_name, guest_phone, guest_count]
    BookingUpdateRequest:
      type: object
      properties:
        room_id:
          type: integer
        check_in:
          type: string
          format: date
        check_out:
          type: string
          format: date
        guest_name:
          type: string
          minLength: 1
          maxLength: 100
        guest_phone:
          type: string
          maxLength: 30
        guest_email:
          type: string
          format: email
          maxLength: 255
        guest_count:
          type: integer
          minimum: 1
          maximum: 10
        total_price:
          type: number
          minimum: 0
        status:
          type: string
          enum: [pending, confirmed, checked_in, checked_out, cancelled, no_show]
        notes:
          type: string
          maxLength: 1000
        source:
          type: string
          enum: [website, phone, line, ota, walk_in, admin]
        ota_platform:
          type: string
          maxLength: 50
        ota_booking_id:
          type: string
          maxLength: 100
      minProperties: 1
    BookingNote:
      type: object
      properties:
        id:
          type: integer
          example: 1
        booking_id:
          type: integer
          example: 101
        content:
          type: string
          example: 旅客來電確認入住時間為下午3點
        created_at:
          type: string
          format: date-time
      required: [id, booking_id, content, created_at]
    RoomSummary:
      type: object
      description: Minimal room info embedded in booking responses
      properties:
        id:
          type: integer
        slug:
          type: string
        name_zh:
          type: string
        type:
          type: string
      required: [id, slug, name_zh, type]
    CalendarEntry:
      type: object
      description: A single day's data for one room in the calendar view
      properties:
        date:
          type: string
          format: date
          example: '2026-07-15'
        room_id:
          type: integer
          example: 1
        room_slug:
          type: string
          example: r1
        room_name_zh:
          type: string
          example: 浪漫情懷 景觀二人房
        status:
          type: string
          enum: [available, booked, blocked, checked_in]
          example: booked
          description: |
            - available: room is free
            - booked: confirmed booking exists
            - blocked: admin blocked this date
            - checked_in: guest currently staying
        booking:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/CalendarBookingSummary'
          description: Booking info if status is booked/checked_in
        blocked_info:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/BlockedDateInfo'
          description: Block info if status is blocked (with reason for hover tooltip)
      required: [date, room_id, room_slug, room_name_zh, status]
    CalendarBookingSummary:
      type: object
      description: Minimal booking info for calendar cells
      properties:
        id:
          type: integer
        guest_name:
          type: string
        guest_count:
          type: integer
        check_in:
          type: string
          format: date
        check_out:
          type: string
          format: date
        status:
          type: string
          enum: [pending, confirmed, checked_in, checked_out, cancelled, no_show]
        source:
          type: string
      required: [id, guest_name, guest_count, check_in, check_out, status, source]
    BlockedDateInfo:
      type: object
      description: Info about a blocked date (shown on hover tooltip)
      properties:
        reason:
          type: string
          example: 老闆出國旅遊
          description: Human-readable reason for blocking (displayed on hover)
        blocked_by:
          type: string
          example: yenfeng
      required: [reason]

    # ── Blocked Dates ──────────────────────────────────────────────────────────
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        per_page:
          type: integer
          minimum: 1
          maximum: 100
          example: 20
        total:
          type: integer
          example: 47
        total_pages:
          type: integer
          example: 3
      required: [page, per_page, total, total_pages]
...
    ApiError:
      type: object
      description: Standard error response
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
              description: |
                Machine-readable error code. Examples:
                - AUTH_INVALID_CREDENTIALS
                - AUTH_SESSION_EXPIRED
                - AUTH_CSRF_INVALID
                - VALIDATION_ERROR
                - NOT_FOUND
                - RATE_LIMITED
                - CONFLICT (e.g. double-booking)
                - INTERNAL_ERROR
                - UPLOAD_FILE_TOO_LARGE
                - UPLOAD_INVALID_TYPE
            message:
              type: string
              example: 驗證失敗
              description: Human-readable error message (Chinese)
            message_en:
              type: string
              example: Validation failed
              description: English fallback message
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    example: check_in
                  message:
                    type: string
                    example: 入住日期不得早於今天
              description: Per-field validation errors (if applicable)
          required: [code, message]
      required: [success, error]
...
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page
...
    BookingIdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      description: Booking ID
...
    DateFromParam:
      name: from
      in: query
      required: true
      schema:
        type: string
        format: date
      description: Start date (inclusive)
    DateToParam:
      name: to
      in: query
      required: true
      schema:
        type: string
        format: date
      description: End date (inclusive)

  # ─── Reusable Responses ──────────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Authentication required or session expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            error:
              code: AUTH_SESSION_EXPIRED
              message: 登入已過期，請重新登入
              message_en: Session expired, please log in again
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: 找不到資源
              message_en: Resource not found
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: 資料驗證失敗
              message_en: Validation failed
              details:
                - field: check_in
                  message: 入住日期不得早於今天
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            error:
              code: RATE_LIMITED
              message: 請求太頻繁，請稍後再試
              message_en: Too many requests, please try again later
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
...
paths:
...
  /bookings:
    get:
      tags: [Bookings]
      operationId: listBookings
      summary: List bookings (admin)
      description: |
        List all bookings with filtering and pagination.
        Admin only.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, checked_in, checked_out, cancelled, no_show]
          description: Filter by booking status
        - name: room_id
          in: query
          schema:
            type: integer
          description: Filter by room
        - name: from
          in: query
          schema:
            type: string
            format: date
          description: Filter check-in date from (inclusive)
        - name: to
          in: query
          schema:
            type: string
            format: date
          description: Filter check-in date to (inclusive)
        - name: search
          in: query
          schema:
            type: string
            maxLength: 100
          description: Search by guest name or phone
        - name: source
          in: query
          schema:
            type: string
            enum: [website, phone, line, ota, walk_in, admin]
          description: Filter by booking source
        - name: sort
          in: query
          schema:
            type: string
            enum: [check_in_asc, check_in_desc, created_at_asc, created_at_desc]
            default: check_in_desc
      responses:
        '200':
          description: Paginated booking list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Booking'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
                required: [success, data, meta]
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Bookings]
      operationId: createBooking
      summary: Create new booking
      description: |
        Public endpoint — used by the website booking form.
        Booking status defaults to "pending".
        Validates room availability (no double-booking or blocked dates).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingCreateRequest'
      responses:
        '201':
          description: Booking created (pending confirmation)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Booking'
                required: [success, data]
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Room not available for requested dates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                success: false
                error:
                  code: CONFLICT
                  message: 所選日期已有預約或不可預訂
                  message_en: Room is not available for the selected dates
        '429':
          $ref: '#/components/responses/RateLimited'
  /bookings/{id}:
    parameters:
      - $ref: '#/components/parameters/BookingIdParam'

    get:
      tags: [Bookings]
      operationId: getBooking
      summary: Get booking detail (admin)
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Booking detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Booking'
                required: [success, data]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Bookings]
      operationId: updateBooking
      summary: Update booking (admin)
      description: |
        Update booking details, status, notes, etc.
        When changing dates/room, validates availability.
      security:
        - cookieAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingUpdateRequest'
      responses:
        '200':
          description: Booking updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Booking'
                required: [success, data]
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Conflict (e.g. double-booking when changing dates)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

    delete:
      tags: [Bookings]
      operationId: deleteBooking
      summary: Delete booking (admin)
      description: Permanently deletes a booking. Use status update to "cancelled" for soft cancel.
      security:
        - cookieAuth: []
          csrfToken: []
      responses:
        '200':
          description: Booking deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: 預約已刪除
                required: [success, data]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /bookings/{id}/notes:
    parameters:
      - $ref: '#/components/parameters/BookingIdParam'
    post:
      tags: [Bookings]
      operationId: addBookingNote
      summary: Add internal note to booking
      description: |
        Append a timestamped internal note to a booking.
        Useful for tracking communications, special requests, etc.
      security:
        - cookieAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 1000
                  example: 旅客來電確認入住時間為下午3點
              required: [content]
      responses:
        '201':
          description: Note added
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/BookingNote'
                required: [success, data]
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /bookings/calendar:
    get:
      tags: [Bookings]
      operationId: getBookingCalendar
      summary: Calendar view of bookings
      description: |
        Returns booking data organized for calendar display.
        Each entry represents one room-day with its status
        (available / booked / blocked / checked_in).

        Blocked dates include a reason/comment that can be shown
        as a tooltip on hover in the admin calendar UI.

        Typically used with a 1-month date range.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/DateFromParam'
        - $ref: '#/components/parameters/DateToParam'
        - name: room_id
          in: query
          schema:
            type: integer
          description: Filter to specific room (omit for all rooms)
      responses:
        '200':
          description: Calendar data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      from:
                        type: string
                        format: date
                      to:
                        type: string
                        format: date
                      rooms:
                        type: array
                        items:
                          type: object
                          properties:
                            room_id:
                              type: integer
                            room_slug:
                              type: string
                            room_name_zh:
                              type: string
                            days:
                              type: array
                              items:
                                $ref: '#/components/schemas/CalendarEntry'
                          required: [room_id, room_slug, room_name_zh, days]
                    required: [from, to, rooms]
                required: [success, data]
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ═══════════════════════════════════════════════════════════════════════════
  # BLOCKED DATES
  # ═══════════════════════════════════════════════════════════════════════════
...